Split itx react hooks from provider for fast-refresh lint
